Move v-next's config to its own folder
