+++ .github/workflows/v-next-ci.yml
@@ -6,6 +6,9 @@ on:
   pull_request:
     branches:
       - "**"
+    paths:
+      - "v-next/**"
+      - "config-v-next/**"
   workflow_dispatch:
 
 concurrency:
