+++ v-next/hardhat-utils/.eslintrc.cjs
@@ -1,3 +1,3 @@
-const { createConfig } = require("../../config/v-next/eslint.cjs");
+const { createConfig } = require("../../config-v-next/eslint.cjs");
 
 module.exports = createConfig(__filename, ["src/index.ts"]);
