+++ v-next/template-package/.eslintrc.cjs
@@ -1,4 +1,4 @@
-const { createConfig } = require("../../config/v-next/eslint.cjs");
+const { createConfig } = require("../../config-v-next/eslint.cjs");
 
 module.exports = createConfig(__filename, [
   "src/index.ts",
